remove renamed files
